requirements.txt
flask
pymongo
python-dotenv

langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
langchain-groq

chromadb
sentence-transformers